+++ src/content/hooks/use-default-branch.test.tsx
@@ -1,33 +1,33 @@
 import { setTimeout } from "node:timers/promises"
 import { renderHook, act } from "@testing-library/react"
 import { describe, it, vi, expect } from "vitest"
-import { useWorkflowFiles } from "@/content/hooks/use-workflow-files"
+import type { BranchSearchResult } from "@/content/repository/branch"
+import { useDefaultBranch } from "@/content/hooks/use-default-branch"
 import { MockRepositoryProvider } from "@/content/repository"
 
-describe("useWorkflowFiles", () => {
+const response = {
+  payload: {
+    branches: {
+      default: {
+        name: "main",
+        isDefault: true,
+        author: {
+          name: "d-kimuson",
+        },
+      },
+    },
+  },
+} as const satisfies BranchSearchResult
+
+describe("useDefaultBranch", () => {
   const repo = { owner: "d-kimuson", repo: "github-actions-search" }
 
   it("呼び出し時は loading 状態を返すこと", () => {
-    const { result } = renderHook(() => useWorkflowFiles(repo), {
+    const { result } = renderHook(() => useDefaultBranch(repo), {
       wrapper: ({ children }) => (
         <MockRepositoryProvider
           repository={{
-            fetchBranches: vi.fn(() =>
-              Promise.resolve({
-                payload: {
-                  branches: {
-                    default: {
-                      name: "main",
-                      isDefault: true,
-                      author: {
-                        name: "d-kimuson",
-                      },
-                    },
-                  },
-                },
-              })
-            ),
-            fetchWorkflowFiles: vi.fn(() => Promise.resolve(["sample.yaml"])),
+            fetchBranches: vi.fn(() => Promise.resolve(response)),
           }}
         >
           {children}
@@ -38,16 +38,16 @@ describe("useWorkflowFiles", () => {
     expect(result.current).toStrictEqual({
       error: undefined,
       loading: true,
-      workflowFiles: undefined,
+      defaultBranch: undefined,
     })
   })
 
   it("レスポンスを受け取った後なら loaded 状態を返すこと", async () => {
-    const { result } = renderHook(() => useWorkflowFiles(repo), {
+    const { result } = renderHook(() => useDefaultBranch(repo), {
       wrapper: ({ children }) => (
         <MockRepositoryProvider
           repository={{
-            fetchWorkflowFiles: vi.fn(() => Promise.resolve(["sample.yaml"])),
+            fetchBranches: vi.fn(() => Promise.resolve(response)),
           }}
         >
           {children}
@@ -62,7 +62,7 @@ describe("useWorkflowFiles", () => {
     expect(result.current).toStrictEqual({
       error: undefined,
       loading: false,
-      workflowFiles: ["sample.yaml"],
+      defaultBranch: "main",
     })
   })
 })