+++ src/content/repository/index.tsx
@@ -1,9 +1,11 @@
 import { createContext, useContext } from "react"
 import type { FC, PropsWithChildren } from "react"
+import { fetchBranches } from "@/content/repository/branch"
 import { fetchWorkflowFiles } from "@/content/repository/workflow"
 
 const defaultRepository = {
   fetchWorkflowFiles,
+  fetchBranches,
 } as const
 
 export type Repository = typeof defaultRepository